refactor(Files hirarchy): Little flatter
