+++ src/lib/request-handlers/bathingspots/get.ts
@@ -1,43 +1,57 @@
-import { getCustomRepository, getRepository } from 'typeorm';
+
 import { Bathingspot } from '../../../orm/entity/Bathingspot';
+import { getCustomRepository } from 'typeorm';
 import { SUCCESS } from '../../messages';
-import { getRegionsList } from '../../repositories/custom-repo-helpers';
+import { getRegionsList, getSpotByUserAndId, getUserWithRelations } from '../../utils/custom-repo-helpers';
 import { RegionRepository } from '../../repositories/RegionRepository';
-import { getResponse, HttpCodes } from '../../types-interfaces';
+import { getResponse, HttpCodes } from '../../common/types-interfaces';
 import { errorResponse, responder, responderWrongId, successResponse } from '../responders';
 import { BathingspotRepository } from '../../repositories/BathingspotRepository';
-
+import { responderWrongIdOrSuccess } from '../responders';
 /**
- * Todo: Which properties should be returned
+ * Gets all the bathingspots of the user
+ * @param request
+ * @param response
  */
-export const getBathingspots: getResponse = async (_request, response) => {
-  let spots: Bathingspot[];
+
+export const getUserBathingspots: getResponse = async (request, response) => {
   try {
-    spots = await getRepository(Bathingspot).find(
-      {
-        where: { isPublic: true },
-      },
-    );
-    responder(response, HttpCodes.success, spots);
-  } catch (e) {
-    response.status(HttpCodes.internalError).json(errorResponse(e));
-  }
-};
-export const getSingleBathingspot: getResponse = async (request, response) => {
-  let spot: Bathingspot | undefined;
-  try {
-    spot = await getRepository(Bathingspot).findOne(request.params.id);
-    if (spot === undefined) {
+    const user = await getUserWithRelations(request.params.userId, ['bathingspots']);
+
+    if (user === undefined) {
       responderWrongId(response);
     } else {
-      responder(response, HttpCodes.success, [spot]);
+      if(user instanceof Bathingspot){
+        throw new Error('Internal Server Error');
+      }
+      responder(response, HttpCodes.success, successResponse(
+        SUCCESS.success200, user.bathingspots));
     }
   } catch (e) {
-    response.status(HttpCodes.internalError).json(errorResponse(e));
+    responder(response, HttpCodes.internalError, errorResponse(e));
   }
 };
 
-export const getBathingspotsByRegion: getResponse = async (request, response) => {
+/**
+ * Gets single bathingspot of user by id
+ * @param request
+ * @param response
+ */
+export const getOneUserBathingspotById: getResponse = async (request, response) => {
+  try {
+    const spotFromUser = await getSpotByUserAndId(request.params.userId, request.params.spotId);
+    responderWrongIdOrSuccess(spotFromUser, response);
+    // if (spotFromUser === undefined) {
+    //   responderWrongId(response);
+    // } else {
+    //   responder(response, HttpCodes.success, successResponse(SUCCESS.success200, [spotFromUser]));
+    // }
+  } catch (e) {
+    responder(response, HttpCodes.internalError, errorResponse(e));
+  }
+};
+
+export const getOneUsersBathingspotsByRegion: getResponse = async (request, response) => {
   try {
     // const regionsRepo = getCustomRepository(RegionRepository);
     // let list = await regionsRepo.getNamesList();
@@ -47,23 +61,19 @@ export const getBathingspotsByRegion: getResponse = async (request, response) =>
     if (!(list.includes(request.params.region))) {
       responderWrongId(response);
     } else {
-      const regionRepo = getCustomRepository(RegionRepository);
       const spotRepo = getCustomRepository(BathingspotRepository);
+      const regionRepo = getCustomRepository(RegionRepository);
       const region = await regionRepo.findByName(request.params.region);
-      let spots: []|any = [];
-      if (region !== undefined) {
-        spots = await spotRepo.findByRegionId(region.id);
-        if (spots === undefined) {
-          spots = [];
+      const userId = request.params.userId;
+      const spots = await spotRepo.findByUserAndRegion(userId, region!.id);
+      if (spots !== undefined) {
+          responder(response, HttpCodes.success, successResponse(SUCCESS.success200, spots));
         } else {
-          spots = spots.filter((spot: Bathingspot) => spot.isPublic === true );
+          responder(response, HttpCodes.success, successResponse(SUCCESS.success200, []));
+
         }
-        responder(response, HttpCodes.success, successResponse(SUCCESS.success200, spots));
-      } else {
-        responderWrongId(response);
-      }
     }
   } catch (e) {
-    response.status(HttpCodes.internalError).json(errorResponse(e));
+    responder(response, HttpCodes.internalError, errorResponse(e));
   }
 };
