+++ src/lib/request-handlers/bathingspots/index.ts
@@ -1,1 +1,5 @@
-export { getBathingspots, getSingleBathingspot } from './get';
+export { getBathingspots, getSingleBathingspot } from './public-get';
+export { deleteBathingspotOfUser } from './delete';
+export { getOneUserBathingspotById, getUserBathingspots } from './get';
+export {addBathingspotToUser} from './post';
+export {updateBathingspotOfUser} from './put';
