+++ src/lib/request-handlers/bathingspots/post.ts
@@ -1,17 +1,17 @@
 import { getManager } from 'typeorm';
-import { User } from '../../../../orm/entity/User';
-import { getUserWithRelations } from '../../../repositories/custom-repo-helpers';
-import { HttpCodes, postResponse, UserRole } from '../../../types-interfaces';
-import { getEntityFields, getMatchingValues } from '../../../utils';
+import { User } from '../../../orm/entity/User';
+import { getUserWithRelations } from '../../utils/custom-repo-helpers';
+import { HttpCodes, postResponse, UserRole } from '../../common/types-interfaces';
+import { getEntityFields, getMatchingValues } from '../../utils';
 import {
   errorResponse,
   responder,
   responderNotAuthorized,
   responderWrongId,
   successResponse,
-} from '../../responders';
+} from '../responders';
 
-import { createSpotWithValues } from './../../../utils/bathingspot-helpers';
+import { createSpotWithValues } from '../../utils/bathingspot-helpers';
 
 
 // const verifyPublic: (obj: any) => boolean = (obj) => {
