--- src/lib/request-handlers/users/bathingspots/index.ts
@@ -1,4 +1,0 @@
-export { deleteBathingspotOfUser } from './delete';
-export { getOneUserBathingspotById, getUserBathingspots } from './get';
-export {addBathingspotToUser} from './post';
-export {updateBathingspotOfUser} from './put';
