+++ src/lib/request-handlers/users/get.ts
@@ -1,7 +1,7 @@
 import { getRepository } from 'typeorm';
 import { User } from '../../../orm/entity/User';
 import { SUCCESS } from '../../messages';
-import { getResponse, HttpCodes } from '../../types-interfaces';
+import { getResponse, HttpCodes } from '../../common/types-interfaces';
 import { errorResponse, responder, responderWrongIdOrSuccess, successResponse } from '../responders';
 
 //  ██████╗ ███████╗████████╗
