+++ src/lib/request-handlers/users/put.ts
@@ -2,11 +2,11 @@ import { getCustomRepository, getRepository } from 'typeorm';
 import { Region } from '../../../orm/entity/Region';
 import { User } from '../../../orm/entity/User';
 import { SUCCESS } from '../../messages';
-import { getRegionsList } from '../../repositories/custom-repo-helpers';
+import { getRegionsList } from '../../utils/custom-repo-helpers';
 import { RegionRepository } from '../../repositories/RegionRepository';
-import { HttpCodes, putResponse } from '../../types-interfaces';
+import { HttpCodes, putResponse } from '../../common/types-interfaces';
 import { errorResponse, responder, responderWrongId, successResponse } from '../responders';
-import { RegionExsists } from './../../types-interfaces';
+import { RegionExsists } from '../../common/types-interfaces';
 
 // ██████╗ ██╗   ██╗████████╗
 // ██╔══██╗██║   ██║╚══██╔══╝
