+++ src/lib/routes.ts
@@ -1,5 +1,4 @@
 import { defaultGetResponse } from './request-handlers/defaults';
-import { postPrediction } from './request-handlers/users/bathingspots/prediction/post';
 import jwtAuthz from 'express-jwt-authz';
 import Router from 'express-promise-router';
 import { checkJwt } from './auth';
@@ -7,7 +6,7 @@ import {
   getBathingspots,
   getSingleBathingspot,
 } from './request-handlers/bathingspots';
-import { getBathingspotsByRegion } from './request-handlers/bathingspots/get';
+import { getBathingspotsByRegion } from './request-handlers/bathingspots/public-get';
 import { deleteRegion, getAllRegions, postRegion, putRegion } from './request-handlers/regions';
 import { getRegionById } from './request-handlers/regions/index';
 // import { defaultGetResponse, defaultPostResponse } from './request-handlers/default-requests';
@@ -24,9 +23,9 @@ import {
   getOneUserBathingspotById,
   getUserBathingspots,
   updateBathingspotOfUser,
-} from './request-handlers/users/bathingspots/';
-import { getOneUsersBathingspotsByRegion } from './request-handlers/users/bathingspots/get';
-import { getCollection, postCollection } from './request-handlers/users/bathingspots/collections';
+} from './request-handlers/bathingspots/';
+import { getOneUsersBathingspotsByRegion } from './request-handlers/bathingspots/get';
+import { getCollection, postCollection } from './request-handlers/bathingspots/collections';
 // import { getPredictions } from './request-handlers/users/bathingspots/prediction/get';
 
 
@@ -64,7 +63,7 @@ router.delete('/users/:userId([0-9]+)/bathingspots/:spotId([0-9]+)',  checkJwt, 
 
 // POST and GET predictions from/to spot
 
-router.post('/users/:userId([0-9]+)/bathingspots/:spotId([0-9]+)/predictions', checkJwt, checkScopes, postPrediction);
+// router.post('/users/:userId([0-9]+)/bathingspots/:spotId([0-9]+)/predictions', checkJwt, checkScopes, postPrediction);
 
 //GET measurements, predictions from
 // router.get('/users/:userId([0-9]+)/bathingspots/:spotId([0-9]+)/predictions', checkJwt, checkScopes, getPredictions);
