+++ src/lib/utils/custom-repo-helpers.ts
@@ -1,10 +1,10 @@
 import { getCustomRepository } from 'typeorm';
 import { Region } from '../../orm/entity/Region';
-import { GetById, GetByIds, GetByIdWithRelations } from '../types-interfaces';
-import { IRegionListEntry } from '../types-interfaces';
-import { BathingspotRepository } from './BathingspotRepository';
-import { RegionRepository } from './RegionRepository';
-import { UserRepository } from './UserRepository';
+import { GetById, GetByIds, GetByIdWithRelations } from '../common/types-interfaces';
+import { IRegionListEntry } from '../common/types-interfaces';
+import { BathingspotRepository } from '../repositories/BathingspotRepository';
+import { RegionRepository } from '../repositories/RegionRepository';
+import { UserRepository } from '../repositories/UserRepository';
 
 export const getUserWithRelations: GetByIdWithRelations = async (userId, relations) => {
   const userRepo = getCustomRepository(UserRepository);
