+++ src/lib/utils/get-geojson-geometry.ts
@@ -1,4 +1,4 @@
-import { getPropsValueGeneric } from '../get-properties-values-generic';
+import { getPropsValueGeneric } from './get-properties-values-generic';
 
 export const getGEOJsonGeometry: (obj: any, key: string) => object = (obj, key) => {
   const geojson = getPropsValueGeneric<{geometry: object}>(obj, key);
