+++ src/orm/entity/index.ts
@@ -7,14 +7,14 @@ export { BathingspotMeasurement } from './BathingspotMeasurement';
 export { BathingspotModel } from './BathingspotModel';
 export { BathingspotPrediction } from './BathingspotPrediction';
 export { BathingspotRawModelData } from './BathingspotRawModelData';
-export { Discharge } from './prediction-measurements/Discharge';
+export { Discharge } from './Discharge';
 export { Event } from './Event';
-export { GenericInput } from './prediction-measurements/generic-input/GenericInput';
-export { GInputMeasurement } from './prediction-measurements/generic-input/GInputMeasurement';
-export { PPlantMeasurement } from './prediction-measurements/pplants/PPlantMeasurement';
-export { PurificationPlant } from './prediction-measurements/pplants/PurificationPlant';
+export { GenericInput } from './GenericInput';
+export { GInputMeasurement } from './GInputMeasurement';
+export { PPlantMeasurement } from './PPlantMeasurement';
+export { PurificationPlant } from './PurificationPlant';
 export { Questionaire } from './Questionaire';
-export { Rain } from './prediction-measurements/Rain';
+export { Rain } from './Rain';
 export { Region } from './Region';
 export { User } from './User';
-export {GlobalIrradiance} from './prediction-measurements/GlobalIrradiance';
+export {GlobalIrradiance} from './GlobalIrradiance';
