+++ src/setup/create-test-user.ts
@@ -1,5 +1,5 @@
-import { UserRole } from '../../lib/types-interfaces';
-import { User } from '../entity/User';
+import { UserRole } from '../lib/common/types-interfaces';
+import { User } from '../orm/entity/User';
 
 // export const createProtectedUser = (): User => {
 //   const protectedUser = new User();
