+++ src/setup/import-existing-data.ts
@@ -2,13 +2,13 @@ import { readdir, readFile, statSync } from 'fs';
 import { extname, resolve } from 'path';
 import proj4 from 'proj4';
 import { promisify } from 'util';
-import { IObject } from '../../lib/types-interfaces';
-import { createSpotWithValues } from '../../lib/utils/bathingspot-helpers';
-import { createMeasurementWithValues } from '../../lib/utils/measurement-helpers';
-import { Bathingspot } from '../entity/Bathingspot';
-import { createPredictionWithValues } from './../../lib/utils/predictions-helpers';
-import { BathingspotMeasurement } from './../entity/BathingspotMeasurement';
-import { BathingspotPrediction } from './../entity/BathingspotPrediction';
+import { IObject } from '../lib/common/types-interfaces';
+import { createSpotWithValues } from '../lib/utils/bathingspot-helpers';
+import { createMeasurementWithValues } from '../lib/utils/measurement-helpers';
+import { Bathingspot } from '../orm/entity/Bathingspot';
+import { createPredictionWithValues } from '../lib/utils/predictions-helpers';
+import { BathingspotMeasurement } from '../orm/entity/BathingspotMeasurement';
+import { BathingspotPrediction } from '../orm/entity/BathingspotPrediction';
 // import {criteria} from '../../lib/utils/bathingspot-helpers';
 const readFileAsync = promisify(readFile);
