+++ src/setup/populate-db.ts
@@ -1,9 +1,9 @@
 import ora = require('ora');
 import { createConnection, getRepository, getConnectionOptions } from 'typeorm';
-import { DefaultRegions, IAddEntitiesToSpotOptions, UserRole } from '../../lib/types-interfaces';
-import { addEntitiesToSpot } from '../../lib/utils/bathingspot-helpers';
-import { Region } from '../entity/Region';
-import { User } from '../entity/User';
+import { DefaultRegions, IAddEntitiesToSpotOptions, UserRole } from '../lib/common/types-interfaces';
+import { addEntitiesToSpot } from '../lib/utils/bathingspot-helpers';
+import { Region } from '../orm/entity/Region';
+import { User } from '../orm/entity/User';
 import { createUser } from './create-test-user';
 import { createMeasurements, createPredictions, createSpots } from './import-existing-data';
